+++ RELEASENOTES.md
@@ -99,6 +99,9 @@
         a custom `VideoFrameProcessor.Factory` during video playback.
     *   Remove `ExoPlayer.setHandleWakeLock(boolean)`, use `setWakeMode(int)`
         instead.
+    *   Remove deprecated
+        `DefaultLoadControl.Builder.createDefaultLoadControl()`, use `build()`
+        instead.
 
 ### 1.0.2 (2023-05-18)